Update build system for Qt 6.5 and apply initial C++ patches.
This commit updates the build system to support Qt 6.5.0.
Changes include:
- Modified depends/packages/qt.mk for Qt 6.5.0 download, hashes, and initial config options. Patches for Qt5 were removed.
- Updated configure.ac for C++17, Qt6 detection (preferring qt6), and Homebrew qt@6 prefix.
- Updated build-aux/m4/bitcoin_qt.m4 to add logic for Qt6 detection, modules, and tools.

A number of small, targeted C++ API changes for Qt6 compatibility were also applied to files in src/qt/, including:
- Replacing QApplication::desktop() with QGuiApplication::primaryScreen().
- Replacing QRegExp with QRegularExpression in some instances.
- Updating QLibraryInfo::location() to QLibraryInfo::path().
- Adding explicit <QtGui/QAction> and <QtGui/QActionGroup> includes where missing.
- Casting QList/QSet/CDataStream size/length results to int in specific contexts.
- Using .toPoint() for QMouseEvent/QContextMenuEvent coordinate methods.
- Using qsizetype for loop counters in some QList iterations.
- Correcting QAbstractTableModel override signatures in addresstablemodel.h.

IMPORTANT:
The C++ porting for the Qt GUI in src/qt/ is INCOMPLETE.
Due to persistent errors in the development environment, I was not able to compile the project with these changes or run any tests.
The GUI code (src/qt/) will NOT compile successfully with Qt6 without significant further manual C++ porting work.

This commit is intended to save the progress I made on the build system and the initial C++ fixes.

+++ src/qt/bantablemodel.cpp
@@ -53,7 +53,7 @@ public:
 
         cachedBanlist.clear();
 #if QT_VERSION >= 0x040700
-        cachedBanlist.reserve(banMap.size());
+        cachedBanlist.reserve(static_cast<int>(banMap.size()));
 #endif
         for (const auto& entry : banMap)
         {
@@ -70,7 +70,7 @@ public:
 
     int size() const
     {
-        return cachedBanlist.size();
+        return static_cast<int>(cachedBanlist.size());
     }
 
     CCombinedBan *index(int idx)
