+++ src/qt/walletmodel.cpp
@@ -257,7 +257,7 @@ WalletModel::SendCoinsReturn WalletModel::prepareTransaction(WalletModelTransact
             total += rcp.amount;
         }
     }
-    if(setAddress.size() != nAddresses)
+    if(static_cast<int>(setAddress.size()) != nAddresses)
     {
         return DuplicateAddress;
     }
@@ -340,7 +340,7 @@ WalletModel::SendCoinsReturn WalletModel::sendCoins(WalletModelTransaction &tran
 
         CDataStream ssTx(SER_NETWORK, PROTOCOL_VERSION);
         ssTx << *newTx->tx;
-        transaction_array.append(&(ssTx[0]), ssTx.size());
+        transaction_array.append(&(ssTx[0]), static_cast<int>(ssTx.size()));
     }
 
     // Add addresses / update labels that we've sent to the address book,
